requirements.txt
temporalio
python-pptx
openai
openpyxl
pandas
streamlit>=1.31.0
